moved get_products from dal products to dal companies
