+++ src/dal/companies.py
@@ -66,3 +66,21 @@ def dal_create_representative(form, owner):
 
         db_conn.users.insert(representative)
         return False
+
+def get_products(name):
+    """Gets products from db
+
+    Will either get all products or based on filter (eg: {'producer': <company-name>})
+
+    Returns:
+        list -- products found in db
+    """
+
+    user = db_conn.users.find_one({'username': name})
+    owner = user['data']['owner']
+
+    products = []
+    for product in db_conn.products.find({'producer': owner}):
+        products.append(str(product))
+
+    return products
